fix: rename folder - placeholder [MDS-1797]
