+++ packages/core/src/placeholder/private/types/PlaceholderProps.ts
